opening animation for launcher
hello,

this is a simple opening animation i made for the launcher. i made this within 15 minutes and i have not received any negative opinions yet

i hope that this animation gets reviewed without much negativity towards this little thing i made, and maybe released to the public

note that this animation can get a little laggy when recording or when using a generally slow computer, but besides that, it should display fine

thanks for reading this message i made, and i look forward to seeing your guys' messages towards this. thanks and god bless.

poopyfart942